Crud viveiros

+++ database/migrations/2024_11_08_113003_create_viveiros_table.php
@@ -15,9 +15,9 @@ return new class extends Migration
             $table->id();
             $table->timestamps();
             $table->string("name");
-            $table->float("width");
-            $table->float("length");
-            $table->float("area");
+            $table->integer("width");
+            $table->integer("length");
+            $table->integer("area");
         });
     }
